Professor working whit system

+++ src/main/java/Entity/Professor.java
@@ -11,10 +11,10 @@ import javax.persistence.Entity;
 @Entity
 @ToString
 public class Professor extends Employee {
-    private Integer allUnit;
+    private Long allUnit;
     private String college;
 
-    public Professor(Integer id, String name, Integer passcode, String status, Integer code, Integer allUnit, String college) {
+    public Professor(Integer id, String name, Integer passcode, String status, Integer code, Long allUnit, String college) {
         super(id, name, passcode, status, code);
         this.allUnit = allUnit;
         this.college = college;
